html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-pointer-move-non-source-into-pointer.html

<!doctype html>
<title>pointer updates (moving non-source nodes into the pointer)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<div id="holder"></div>
<script>
var brEvents = [];
var textEvents = [];
var a1 = 0;
var b1 = 0;
var c1 = 0;
var a2 = 0;
var b2 = 0;
var c2 = 0;
</script>
<video id="v-br"><source id="br-a" onerror="a1++; brEvents.push('a')"><source id="br-b" onerror="b1++; brEvents.push('b')" src="resources/delayed-broken-video.py?br"><source id="br-c" onerror="c1++; brEvents.push('c')"></video>
<video id="v-text"><source id="text-a" onerror="a2++; textEvents.push('a')"><source id="text-b" onerror="b2++; textEvents.push('b')" src="resources/delayed-broken-video.py?text"><source id="text-c" onerror="c2++; textEvents.push('c')"></video>
<script>
async_test(function(t) {
  var holder = document.getElementById('holder');
  var br = document.createElement('br');
  holder.appendChild(br);
  document.getElementById('v-br').insertBefore(br, document.getElementById('br-c'));

  var text = document.createTextNode('x');
  holder.appendChild(text);
  document.getElementById('v-text').insertBefore(text, document.getElementById('text-c'));

  window.onload = t.step_func(function() {
    assert_equals(a1, 1, 'error events on br test a');
    assert_equals(b1, 1, 'error events on br test b');
    assert_equals(c1, 1, 'error events on br test c');
    assert_array_equals(brEvents, ['a', 'b', 'c'], 'br test processed source order');

    assert_equals(a2, 1, 'error events on text test a');
    assert_equals(b2, 1, 'error events on text test b');
    assert_equals(c2, 1, 'error events on text test c');
    assert_array_equals(textEvents, ['a', 'b', 'c'], 'text test processed source order');
    t.done();
  });
});
</script>
